pyproject.toml
[tool.poetry]
name = "Ryoma"
version = "0.1.0"
description = "R Agent Project"
authors = ["Edison Zhang <zhangn661@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
langgraph = "^0.0.10"
langchain_aws = "*"
langchain_openai = "*"
boto3 = "^1.34.34"
langchain-google-genai = "^0.0.11"
fastapi = "^0.110.0"
uvicorn = "^0.30.6"
click = "^8.1.7"
pydantic-settings = "^2.1.0"
colorama = "^0.4.6"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.9.0"
pytest = "^8.1.1"
ipython = "^8.22.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
